minicpm-livekit-agent/pyproject.toml
[project]
name = "livekit-qwen-agent"
version = "0.1.0"
description = "LiveKit voice agent powered by Qwen-Omni Realtime (DashScope)"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-openai",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
]

[project.scripts]
agent = "agent:main"
